feat(admin-ui): clean up, rename states, move new feature to component :)
